Conflicts:
	rts/CMakeLists.txt
	rts/Rendering/Models/3DModel.cpp
	rts/Rendering/Models/3DModel.h
	rts/Rendering/Models/IModelParser.cpp
	rts/Rendering/Models/IModelParser.h
	rts/lib/CMakeLists.txt

+++ installer/make_source_package.sh
@@ -1,12 +1,13 @@
 #!/bin/sh
-# Author: Tobi Vollebregt
 
 # Quit on error.
 set -e
 
+ORIG_DIR=$(pwd)
+
 # Sanity check.
 if [ ! -x /usr/bin/git ]; then
-	echo "Error: Couldn't find /usr/bin/git"
+	echo "Error: Couldn't find /usr/bin/git" >&2
 	exit 1
 fi
 
@@ -14,12 +15,12 @@ fi
 # (Compatible with SConstruct, which is in trunk root)
 
 while [ ! -d installer ]; do
-        if [ "$PWD" = "/" ]; then
-                echo "Error: Could not find installer directory."
-                echo "Make sure to run this script from a directory below your checkout directory."
-                exit 1
-        fi
-        cd ..
+	if [ "$PWD" = "/" ]; then
+		echo "Error: Could not find installer directory." >&2
+		echo "Make sure to run this script from a directory below your checkout directory." >&2
+		exit 1
+	fi
+	cd ..
 done
 
 set +e # turn of quit on error
@@ -40,14 +41,13 @@ else
 	version_string=`git describe --tags | sed s/\-[^\-]*$//`
 	branch="master"
 fi
-echo "Using $branch as source"
+echo "Using ${branch} as source"
 
 dir="spring_${version_string}"
 
 # Each one of these that is set, is built when running this script.
 # Linux archives
 # * linux (LF) line endings
-# * removed files needed for windows installer generation only
 # * GPL compatible
 lzma="spring_${version_string}_src.tar.lzma"
 #tbz="spring_${version_string}_src.tar.bz2"
@@ -63,9 +63,9 @@ tgz="spring_${branch}_src.tar.gz"
 # (directories are included recursively)
 include=" \
  $dir/AI/ \
- $dir/Documentation/ \
+ $dir/doc/ \
  $dir/Doxyfile \
- $dir/game/ \
+ $dir/cont/ \
  $dir/installer/ \
  $dir/LICENSE.html \
  $dir/README.* \
@@ -74,24 +74,12 @@ include=" \
  $dir/tools/SelectionEditor/ \
  $dir/CMakeLists.txt \
  $dir/tools/unitsync/ \
+ $dir/tools/DemoTool/ \
  $dir/tools/DedicatedServer/"
 
 # On linux, win32 executables are useless.
-# TASClient is windows only.
 exclude_from_all=""
-linux_exclude="${exclude_from_all}
-	${dir}/installer/include/
-	${dir}/installer/sections/
-	${dir}/installer/graphics/
-	${dir}/installer/nsis_plugins/
-	${dir}/installer/*.exe
-	${dir}/installer/*.bat
-	${dir}/installer/*.nsi
-	${dir}/installer/*.nsh
-	${dir}/installer/make_installer.pl
-	${dir}/installer/make_luaui_nsh.py
-	${dir}/installer/tasclient_download.sh
-	${dir}/installer/springlobby_download.sh"
+linux_exclude="${exclude_from_all}"
 linux_include=""
 windows_exclude="${exclude_from_all}"
 windows_include=""
@@ -105,31 +93,34 @@ git submodule update --init
 cd ..
 [ -n "$linux_exclude" ] && rm -rf $linux_exclude
 [ -n "$lzma" ] && echo "Creating .tar.lzma archive ($lzma)" && \
-	tar --lzma -c -f "../$lzma" $include $linux_include
+	tar --lzma -c -f "../$lzma" $include $linux_include --exclude=.git
 [ -n "$tbz" ] && echo "Creating .tar.bz2 archive ($tbz)" && \
-	tar -c -j -f "../$tbz" $include $linux_include
+	tar -c -j -f "../$tbz" $include $linux_include --exclude=.git
 [ -n "$tgz" ] && echo "Creating .tar.gz archive ($tgz)" && \
-	tar -c -z -f "../$tgz" $include $linux_include
+	tar -c -z -f "../$tgz" $include $linux_include --exclude=.git
 cd ..
 echo 'Cleaning'
 rm -rf lf
 
-### TODO: needs fixing (not really using CRLF)
-# Windows line endings, .zip/.7z package
-#echo 'Exporting checkout dir with CRLF line endings'
-git clone -n . crlf/$dir
-cd crlf/$dir
-git config core.autocrlf true
-git checkout $branch
-git submodule update --init
-cd ..
+if [ -n "$zip" ] || [ -n "$seven_zip" ]; then
+	### TODO: needs fixing (not really using CRLF)
+	# Windows line endings, .zip/.7z package
+	#echo 'Exporting checkout dir with CRLF line endings'
+	git clone -n . crlf/$dir
+	cd crlf/$dir
+	git config core.autocrlf true
+	git checkout $branch
+	git submodule update --init
+	cd ..
 
-[ -n "$windows_exclude" ] && rm -rf $windows_exclude
-[ -n "$zip" ] && [ -x /usr/bin/zip ] && echo "Creating .zip archive ($zip)" && \
-	/usr/bin/zip -q -r -u -9 "../$zip" $include $windows_include
-[ -n "$seven_zip" ] && [ -x /usr/bin/7z ] && echo "Creating .7z archive ($seven_zip)" && \
-	/usr/bin/7z a -t7z -m0=lzma -mx=9 -mfb=64 -md=32m -ms=on "../$seven_zip" $include >/dev/null
-cd ..
-echo 'Cleaning'
-rm -rf crlf
-cd ..
+	[ -n "$windows_exclude" ] && rm -rf $windows_exclude
+	[ -n "$zip" ] && [ -x /usr/bin/zip ] && echo "Creating .zip archive ($zip)" && \
+		/usr/bin/zip -q -r -u -9 "../$zip" $include $windows_include
+	[ -n "$seven_zip" ] && [ -x /usr/bin/7z ] && echo "Creating .7z archive ($seven_zip)" && \
+		/usr/bin/7z a -t7z -m0=lzma -mx=9 -mfb=64 -md=32m -ms=on "../$seven_zip" $include >/dev/null
+	cd ..
+	echo 'Cleaning'
+	rm -rf crlf
+fi
+
+cd ${ORIG_DIR}
